fix: cast eval_match win rate to Python float
numpy float64 leaking into best_pen made saved checkpoints fail
torch.load(weights_only=True).

+++ training/train_ppo.py
@@ -272,7 +272,7 @@ def eval_match(
             opp_total += float(totals[1:].mean())
     net.train()
     return {
-        "match_win": won / matches,
+        "match_win": float(won / matches),
         "policy_total": pol_total / matches,
         "opp_total": opp_total / matches,
     }
